Move main.py to root directory
